Refactor the topLevel documentation react components for 0x.js and Smart contracts into a single component
